Update app id to "com.dono.android"
