+++ src/app/build.gradle
@@ -17,7 +17,7 @@ android {
     buildToolsVersion "23.0.2"
 
     defaultConfig {
-        applicationId "com.dono.psakkos.dono"
+        applicationId "com.dono.android"
         minSdkVersion 21
         targetSdkVersion 23
         versionCode 14
