+++ src/app/src/main/java/com/dono/android/DonoToastFactory.java
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package com.dono.psakkos.dono;
+package com.dono.android;
 
 import android.content.Context;
 import android.content.res.Resources;
