+++ src/app/src/main/java/com/dono/android/fragments/EmptyFragment.java
@@ -14,25 +14,24 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package com.dono.psakkos.dono;
+package com.dono.android.fragments;
 
-import com.dono.psakkos.dono.core.Dono;
+import android.app.Fragment;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
-import org.junit.Test;
+import com.dono.android.R;
 
-import static org.junit.Assert.*;
-
-public class DonoCoreConstantsTests
+public class EmptyFragment extends Fragment
 {
-    @Test
-    public void requireThatMinKeyLengthIs17() throws Exception
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState)
     {
-        assertEquals(17, Dono.MIN_KEY_LENGTH);
-    }
-
-    @Test
-    public void requireThatMaxDkLenIs32() throws Exception
-    {
-        assertEquals(64, Dono.MAX_DK_LEN);
+        View view = inflater.inflate(R.layout.empty_fragment, container, false);
+        return view;
     }
 }
