+++ src/app/src/main/java/com/dono/android/fragments/KeyFragment.java
@@ -14,9 +14,8 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package com.dono.psakkos.dono.fragments;
+package com.dono.android.fragments;
 
-import android.app.Fragment;
 import android.content.Context;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -30,10 +29,9 @@ import android.widget.EditText;
 import android.widget.ImageButton;
 import android.widget.LinearLayout;
 
-import com.dono.psakkos.dono.core.Dono;
-import com.dono.psakkos.dono.MainActivity;
-import com.dono.psakkos.dono.core.PersistableKey;
-import com.dono.psakkos.dono.R;
+import com.dono.android.core.Dono;
+import com.dono.android.core.PersistableKey;
+import com.dono.android.R;
 
 public class KeyFragment extends DonoFragment
 {
