+++ src/app/src/main/res/layout/content_main.xml
@@ -9,7 +9,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="com.dono.psakkos.dono.MainActivity"
+    tools:context="com.dono.android.MainActivity"
     tools:showIn="@layout/app_bar_main">
     
 </RelativeLayout>
