+++ src/app/src/test/java/com/dono/android/DonoCoreKeyDerivationUnitTests.java
@@ -14,9 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-package com.dono.psakkos.dono;
+package com.dono.android;
 
-import com.dono.psakkos.dono.core.Dono;
+import com.dono.android.core.Dono;
 
 import org.junit.Test;
